Fix all of the Core passes for Cursors
Lift, Strictness and Normalize had to be changed so that everything in
Iridium works correctly. PackedToCursor now also works differently. Note
that the Has cursors have not yet been implemented, and the FinishCursor
function now returns an invalid PACKED_Int. This needs the newest
version of lvm (from the branch with the same name) to work as well!