Change how cursors work, write lengths of fields
Cursor types have been changed in accordance to Lvm.

The main thing that has changed apart from that, is that cursors now
write the lengths of their fields before writing the fields themselves.
The exact way this works might have to be changed later, as this is done
in the WriteCursor call itself, and as such, is very inflexible.