Change cursor translation to be more intelligent
In theory, addCursorsToExpr should be intelligent enough now to be able
to handle multi-parameter constructors. However, in practice, it doesn't
work yet, because other parts of this translation are not intelligent
enough for this yet.

PackedToCursor now also uses the new pattern synonyms that are in
Lvm.Core.Type, which makes some things _much_ more legible.

The change to LLVM.Env is important, but I feel like it's supposed to be
part of a previous commit. It's required for some step in the
translation process now, since LLVM needs to know the types of
constructors (because of prim_write_ctor).