Properly read cursor fields in multi-arg ctors
The function compileExtractCursorFields used to crash under almost any
circumstance when reading from any field but the first in a constructor
with multiple arguments. It has now been changed to read all arguments,
first and foremost, and also to skip any arguments that do not have
names. The combination of this should skip both reasons for crashes.