Make writing nested PACKED_ types work
It doesn't even work completely. The problem is that we need to know
which constructor has been written before it has been written. This is
clearly not a very easy task.

The fix is to use a single cursor to write, rather than make a new
cursor for the inner cursor writes.