Mostly fix PackedToCursor
It works for many function calls to lifted constructors now. However,
the problem is, that it basically only works for the one case statement
that the single test consisted of so far. It also doesn't work whenever
the constructor is an argument to another function. These two things
should be focused on next.