Final commit for adding Cursors
Adding cursors works now! However, it reduces performance overall. Read
the report for more information on that.

There are some other small changes, but the changes in the Haskell code
have explanatory comments.
The changes in CursorFunctions are essentially there to remove debugging
code. These calls to the cursor memdump are in comments now, so it is
easy to reactivate them.
The change to the helium.cabal file is small, but important, as it adds
the CursorFunctions.ll file to the required files.